browse datasets alphabetically, clean up
