+++ frontend/pages/3_GraphQL.py
@@ -1,11 +1,12 @@
 import streamlit as st
 from streamlit_ace import st_ace
-from utils import (
+from src.queries import browse_datasets_query
+from src.utils import (
     execute_query
 )
 
 st.markdown("# GraphQL")
-# st.sidebar.header("Explore Datasets")
+st.write("For information on how to use GraphQL with Weaviate, please refer to the [documentation](https://weaviate.io/developers/weaviate/api/graphql).")
 st.write("Enter a GraphQL query below:")
 
 query = st_ace(
@@ -13,22 +14,13 @@ query = st_ace(
     theme="twilight",
     keybinding="vscode",
     tab_size=4
+
 )
 
 if query:
     result = execute_query(query)
     data = result['data']['Get']['Dataset']
     st.json(data)
-    # with st.container():
-    #     for dataset in data:
-    #         st.markdown(f"## {dataset['name']}")
-    #         st.markdown(f"{dataset['description']}")
 
-# {
-#     Get {
-#         Dataset(limit: 3) {
-#             name
-#             description
-#         }
-#     }
-# }
+with st.expander("Example"):
+    st.markdown(f"```\n{browse_datasets_query.format()}\n```")
