+++ frontend/src/queries.py
@@ -1,4 +1,4 @@
-explore_datasets_query = \
+search_datasets_query = \
 """
 {{
     Get {{
@@ -50,6 +50,57 @@ explore_datasets_query = \
             }}
             _additional {{
                 distance
+            }}
+        }}
+    }}
+}}
+"""
+
+browse_datasets_query = \
+"""
+{{
+    Get {{
+        Dataset (
+            limit: 1000
+        ) {{
+            name
+            description
+            documentation
+            tags
+            managedBy {{
+                ... on Publisher {{
+                    name
+                }}
+            }}
+            hasResource {{
+                ... on Resource {{
+                    arn
+                    region
+                    description
+                    type
+                    requesterPays
+                }}
+            }}
+            hasTutorial {{
+                ... on Tutorial {{
+                    title
+                    url
+                    services
+                }}
+            }}
+            hasPublication {{
+                ... on Publication {{
+                    title
+                    url
+                    authorName
+                }}
+            }}
+            hasToolOrApplication {{
+                ... on ToolOrApplication {{
+                    title
+                    url
+                    authorName
+                }}   
             }}
         }}
     }}
