+++ frontend/src/utils.py
@@ -1,23 +1,60 @@
+import os
+from dotenv import load_dotenv, find_dotenv
+load_dotenv(find_dotenv());
+import json
 import streamlit as st
+import weaviate
+from src.queries import search_datasets_query, browse_datasets_query
 
-def render_results(datasets, limit):
-    if limit:
-        st.text(f"{len(datasets)} results")
-    if len(datasets) > 0:
-        with st.expander("Raw JSON"):
-            st.json(datasets[0], expanded=True)
+
+@st.cache(allow_output_mutation=True)
+def get_client():
+    return weaviate.Client(os.environ["WEAVIATE_ENDPOINT"])
+
+client = get_client()
+
+def execute_query(query):
+    result = client.query.raw(query)
+    result = json.dumps(result) \
+        .replace('<br>', '') \
+        .replace('<br/>', '') \
+        .replace(' <br />', '')
+    return json.loads(result)
+
+@st.cache(allow_output_mutation=True)
+def run_semantic_search_query(concepts, limit=1000, distance=0.8, search_datasets_query=search_datasets_query):
+
+    search_datasets_query = search_datasets_query.format(
+        concepts=json.dumps(concepts.split(",")), 
+        limit=str(limit),
+        distance=str(distance)
+        )
+
+    return client.query.raw(search_datasets_query)['data']['Get']['Dataset']
+
+@st.cache(allow_output_mutation=True)
+def run_browse_datasets_query():
+    return client.query.raw(browse_datasets_query.format())['data']['Get']['Dataset']
+
+
+def render_results(datasets, limit=None):
+    st.text(f"{len(datasets)} results")
+    # if len(datasets) > 0:
+    #     with st.expander("Raw JSON"):
+    #         st.json(datasets[0], expanded=True)
     for idx, dataset in enumerate(datasets):
         st.markdown("------------")
         st.markdown(f"### ({idx+1}) {dataset['name']}")
-        st.text(f"Distance: {round(dataset['_additional']['distance'], 2)}")
+        if '_additional' in dataset:
+            st.text(f"Distance: {round(dataset['_additional']['distance'], 2)}")
         st.markdown(
             f"{dataset['description'].replace('<br>', '').replace('<br/>', '')}"
             )
         st.markdown(f"**Managed by:**")
         st.markdown('- '.join([ item['name']+'\n' for item in dataset['managedBy'] ]))
         st.markdown(f"**Tags:**")
-        # st.markdown('- '.join([ item['name']+'\n' for item in dataset['hasTag'] ]))
-        st.markdown(f"{dataset['tags']}")
+        tags = ", ".join([tag for tag in dataset['tags'].split(',') if tag != 'aws-pds'])
+        st.markdown(f"{tags}")
         with st.expander("**Documentation**"):
             st.markdown(f"{dataset['documentation']}")
         with st.expander("**Publications**"):
@@ -32,7 +69,7 @@ def render_results(datasets, limit):
                     st.markdown(f"**{resource['type']}**")
                     st.markdown(f"{resource['description']}")
                     if resource['requesterPays']:
-                        st.markdown("(**requester pays**)")
+                        st.markdown("**requester pays**")
                     st.code(resource['arn'], language="bash")
                     st.markdown(f"{resource['region']}")
         with st.expander("**Tutorials**"):
